# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "-",
        "title": "Documenting Whitebook API",
        "contact": {},
        "version": "1"
    },
    "basePath": "/",
    "paths": {
...
        "/cart/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/cart/delete/{cartId}": {
            "delete": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/cart/update": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/cart/{cartId}": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/carts": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/category/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/category/delete/{categoryId}": {
            "delete": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/category/update": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/category/{categoryId}": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}